Separated logging into model related logging and ml_tooling related logging
